Revert "Support AIPS beam defined in history header (#1275)"
This reverts commit f118d7c6dbf8b3d1709fb483581cd9e0de020060.
