+++ CMakeLists.txt
@@ -208,7 +208,6 @@ set(SOURCE_FILES
         src/ImageData/CompressedFits.cc
         src/ImageData/FileInfo.cc
         src/ImageData/FileLoader.cc
-        src/ImageData/FitsLoader.cc
         src/ImageData/Hdf5Attributes.cc
         src/ImageData/Hdf5Loader.cc
         src/ImageData/PolarizationCalculator.cc
