+++ src/FileList/FileExtInfoLoader.cc
@@ -39,7 +39,7 @@ bool FileExtInfoLoader::FillFitsFileInfoMap(
     bool map_ok(false);
 
     if (IsCompressedFits(filename)) {
-        CompressedFits cfits(filename, _loader->GetAipsBeamSupport());
+        CompressedFits cfits(filename);
         if (!cfits.GetFitsHeaderInfo(hdu_info_map)) {
             message = "Compressed FITS headers failed.";
             return map_ok;
@@ -53,6 +53,11 @@ bool FileExtInfoLoader::FillFitsFileInfoMap(
             int spectral_axis = cfits.GetSpectralAxis();
             int stokes_axis = cfits.GetStokesAxis();
             AddComputedEntriesFromHeaders(hdu_info.second, render_axes, spectral_axis, stokes_axis, &cfits);
+
+            const casacore::ImageBeamSet beam_set = cfits.GetBeamSet();
+            if (!beam_set.empty()) {
+                AddBeamEntry(hdu_info.second, beam_set);
+            }
         }
     } else {
         // Get list of image HDUs
@@ -133,7 +138,6 @@ bool FileExtInfoLoader::FillFileInfoFromImage(CARTA::FileInfoExtended& extended_
 
             bool check_data_type(false);
             auto image = _loader->GetImage(check_data_type);
-            bool is_history_beam = _loader->IsHistoryBeam();
 
             if (image) {
                 // Check dimensions
@@ -150,6 +154,7 @@ bool FileExtInfoLoader::FillFileInfoFromImage(CARTA::FileInfoExtended& extended_
                 auto equivalent_type = data_type; // for FITS only, for rescaled data
                 casacore::String image_type(image->imageType());
                 bool use_image_for_entries(false);
+
                 if (image_type == "FITSImage") {
                     // casacore FitsKeywordList has incomplete header names (no n on CRVALn, CDELTn, CROTA, etc.) so read with fitsio
                     casacore::FITSImage* fits_image = dynamic_cast<casacore::FITSImage*>(image.get());
@@ -212,8 +217,7 @@ bool FileExtInfoLoader::FillFileInfoFromImage(CARTA::FileInfoExtended& extended_
                         extended_info, image_shape, spatial_axes, spectral_axis, stokes_axis, render_axes, depth_axis, axes_names);
 
                     // Computed entries for rendered image axes, depth axis (may not be spectral), stokes axis
-                    AddComputedEntries(
-                        extended_info, image.get(), render_axes, spectral_axis, stokes_axis, use_image_for_entries, is_history_beam);
+                    AddComputedEntries(extended_info, image.get(), render_axes, spectral_axis, stokes_axis, use_image_for_entries);
                     info_ok = true;
                 }
             } else { // image failed
@@ -820,7 +824,7 @@ void FileExtInfoLoader::AddShapeEntries(CARTA::FileInfoExtended& extended_info, 
 }
 
 void FileExtInfoLoader::AddComputedEntries(CARTA::FileInfoExtended& extended_info, casacore::ImageInterface<float>* image,
-    const std::vector<int>& display_axes, int spectral_axis, int stokes_axis, bool use_image_for_entries, bool is_history_beam) {
+    const std::vector<int>& display_axes, int spectral_axis, int stokes_axis, bool use_image_for_entries) {
     // Add computed entries to extended file info
     if (use_image_for_entries) {
         // Use image coordinate system
@@ -953,7 +957,7 @@ void FileExtInfoLoader::AddComputedEntries(CARTA::FileInfoExtended& extended_inf
     casacore::ImageInfo image_info = image->imageInfo();
     if (image_info.hasBeam()) {
         const casacore::ImageBeamSet beam_set = image_info.getBeamSet();
-        AddBeamEntry(extended_info, beam_set, is_history_beam);
+        AddBeamEntry(extended_info, beam_set);
     }
 
     AddCoordRanges(extended_info, image->coordinates(), image->shape());
@@ -1254,7 +1258,6 @@ void FileExtInfoLoader::AddComputedEntriesFromHeaders(CARTA::FileInfoExtended& e
                         spectral_crval = rest_freq / (fits_offset + spectral_crval / casacore::C::c);
                         spectral_cdelt = rest_freq / (fits_offset + spectral_cdelt / casacore::C::c);
                     }
-
                     spectral_coordinate =
                         casacore::SpectralCoordinate(frequency_type, spectral_crval, spectral_cdelt, spectral_crpix, rest_freq);
                 }
@@ -1277,22 +1280,11 @@ void FileExtInfoLoader::AddComputedEntriesFromHeaders(CARTA::FileInfoExtended& e
             coordsys.addCoordinate(stokes_coord);
         }
 
-        // Beams read from header, BEAMS table, or history
-        bool is_history_beam(false);
-        const casacore::ImageBeamSet beam_set = compressed_fits->GetBeamSet(is_history_beam);
-        if (!beam_set.empty()) {
-            AddBeamEntry(extended_info, beam_set, is_history_beam);
-            if (is_history_beam) {
-                // For logging
-                _loader->SetHistoryBeam(beam_set.getBeam());
-            }
-        }
-
         AddCoordRanges(extended_info, coordsys, shape);
     }
 }
 
-void FileExtInfoLoader::AddBeamEntry(CARTA::FileInfoExtended& extended_info, const casacore::ImageBeamSet& beam_set, bool is_history_beam) {
+void FileExtInfoLoader::AddBeamEntry(CARTA::FileInfoExtended& extended_info, const casacore::ImageBeamSet& beam_set) {
     // Add restoring/median beam to computed entries.
     casacore::GaussianBeam gaussian_beam;
     std::string entry_name;
@@ -1305,48 +1297,13 @@ void FileExtInfoLoader::AddBeamEntry(CARTA::FileInfoExtended& extended_info, con
     }
 
     if (!gaussian_beam.isNull()) {
-        casacore::Quantity major = gaussian_beam.getMajor();
-        casacore::Quantity minor = gaussian_beam.getMinor();
-        double pa = gaussian_beam.getPA().get("deg").getValue();
+        std::string beam_info = fmt::format("{:g}\" X {:g}\", {:g} deg", gaussian_beam.getMajor("arcsec"), gaussian_beam.getMinor("arcsec"),
+            gaussian_beam.getPA(casacore::Unit("deg")));
 
-        std::string beam_info =
-            fmt::format("{:g}\" X {:g}\", {:g} deg", major.get("arcsec").getValue(), minor.get("arcsec").getValue(), pa);
-
-        if (is_history_beam) {
-            beam_info += " (extracted from HISTORY)";
-        }
-
-        auto computed_entry = extended_info.add_computed_entries();
-        computed_entry->set_name(entry_name);
-        computed_entry->set_entry_type(CARTA::EntryType::STRING);
-        computed_entry->set_value(beam_info);
-
-        if (is_history_beam) {
-            // Append beam header entries in degrees
-            double major_deg = major.get("deg").getValue();
-            double minor_deg = minor.get("deg").getValue();
-
-            auto header_entry = extended_info.add_header_entries();
-            header_entry->set_name("BMAJ");
-            header_entry->set_entry_type(CARTA::EntryType::FLOAT);
-            header_entry->set_value(fmt::format("{:E}", major_deg));
-            header_entry->set_numeric_value(major_deg);
-            header_entry->set_comment("extracted from HISTORY");
-
-            header_entry = extended_info.add_header_entries();
-            header_entry->set_name("BMIN");
-            header_entry->set_entry_type(CARTA::EntryType::FLOAT);
-            header_entry->set_value(fmt::format("{:E}", minor_deg));
-            header_entry->set_numeric_value(minor_deg);
-            header_entry->set_comment("extracted from HISTORY");
-
-            header_entry = extended_info.add_header_entries();
-            header_entry->set_name("BPA");
-            header_entry->set_entry_type(CARTA::EntryType::FLOAT);
-            header_entry->set_value(fmt::format("{:E}", pa));
-            header_entry->set_numeric_value(pa);
-            header_entry->set_comment("extracted from HISTORY");
-        }
+        auto entry = extended_info.add_computed_entries();
+        entry->set_name(entry_name);
+        entry->set_entry_type(CARTA::EntryType::STRING);
+        entry->set_value(beam_info);
     }
 }
 
@@ -1619,8 +1576,7 @@ casacore::Vector<casacore::String> FileExtInfoLoader::FitsHeaderStrings(casacore
     for (int i = 0; i < nkeys; ++i) {
         header_strings[i] = header_str.substr(pos, 80);
 
-        if (!header_strings[i].contains("HISTORY") && header_strings[i].contains("FELO-HEL")) {
-            // not supported, use ImageHeaderToFITS instead
+        if (header_strings[i].contains("FELO-HEL")) {
             header_strings.resize();
             return header_strings;
         }
