+++ src/Util/Casacore.h
@@ -29,7 +29,4 @@ std::string FormatQuantity(const casacore::Quantity& quantity);
 // Convert unit to one recognized by casacore (case-sensitive)
 void NormalizeUnit(casacore::String& unit);
 
-// Parse AIPS beam header using regex_match
-bool ParseHistoryBeamHeader(std::string& header, std::string& bmaj, std::string& bmin, std::string& bpa);
-
 #endif // CARTA_BACKEND__UTIL_CASACORE_H_
