+++ test/TestFileInfo.cc
@@ -186,10 +186,8 @@ public:
         CARTA::FileInfoResponse response;
         auto& file_info = *response.mutable_file_info();
         std::map<std::string, CARTA::FileInfoExtended> extended_info_map;
-        bool support_aips_beam(false);
         string message;
-        bool success = FillExtendedFileInfo(
-            extended_info_map, file_info, request.directory(), request.file(), request.hdu(), support_aips_beam, message);
+        bool success = FillExtendedFileInfo(extended_info_map, file_info, request.directory(), request.file(), request.hdu(), message);
         if (success) {
             *response.mutable_file_info_extended() = {extended_info_map.begin(), extended_info_map.end()};
         }
@@ -245,9 +243,8 @@ TEST_F(FileExtInfoLoaderTest, FitsHistoryEntries) {
     std::string hdu;
     CARTA::FileInfoExtended extended_info;
     CARTA::FileInfo file_info;
-    bool support_aips_beam(false);
-    bool success = t_session.FillExtendedFileInfo(extended_info, file_info, TestRoot() / "data" / "images" / "fits", "noise_10px_10px.fits",
-        hdu, support_aips_beam, message, full_name);
+    bool success = t_session.FillExtendedFileInfo(
+        extended_info, file_info, TestRoot() / "data" / "images" / "fits", "noise_10px_10px.fits", hdu, message, full_name);
     EXPECT_EQ(success, true);
 
     int num_history_entries = 0;
